Add training monitoring: tqdm progress bar, TensorBoard logging, monitoring utilities
Agent-Logs-Url: https://github.com/apehex/deformers/sessions/00134fd4-a100-4656-815f-446cd64f1126

+++ src/deformers/pipelines/eval.py
@@ -45,5 +45,5 @@ def topk_rate(
     """Fraction of (B, T) positions where teacher top-k and student top-k token sequences match exactly."""
     __t = teacher_arr.topk(k_num, dim=-1).indices
     __s = student_arr.topk(k_num, dim=-1).indices
-    # reduced to a scalar
-    return (__t == __s).all(dim=-1).float().mean()
+    # reduced to a Python float
+    return float((__t == __s).all(dim=-1).float().mean())
